Initial Commit
A GMail Backup tool that allows you to save all or selected labels to your computer for storage.  To use this, you must enter client id and client secret from Google Cloud Console/GMail API setup which this does not do.  The actual program can be found at TSTP.xyz and does not require the setup but does require you to be added to a test users list prior to usage or else it will not see you as having access when going through the Google authentication process.

Better readme coming soon.